1. prototype uboot v1.7 2. hdmi stick uboot v2.1 3. pad ref uboot v4.7

+++ fs/ext4/ext4fs.c
@@ -1040,7 +1040,7 @@ int ext4fs_write(const char *fname, unsigned char *buffer,
 	int inodeno;
 	time_t timestamp = 0;
 
-	uint64_t bytes_reqd_for_file;
+	uint32_t bytes_reqd_for_file;
 	unsigned int blks_reqd_for_file;
 	unsigned int blocks_remaining;
 	int existing_file_inodeno;
